+++ test/cases/phase0/math/abs-min-max/expected.c
@@ -3,9 +3,9 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", fabs(-5));
-    printf("%g\n", fabs(3.14));
-    printf("%g\n", fmin(1, 2));
-    printf("%g\n", fmax(1, 2));
+    printf("%g\n", (double)(fabs(-5)));
+    printf("%g\n", (double)(fabs(3.14)));
+    printf("%g\n", (double)(fmin(1, 2)));
+    printf("%g\n", (double)(fmax(1, 2)));
     return 0;
 }
