+++ test/cases/phase0/math/hyperbolic/expected.c
@@ -3,11 +3,11 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", sinh(0.0));
-    printf("%g\n", cosh(0.0));
-    printf("%g\n", tanh(0.0));
-    printf("%g\n", asinh(0.0));
-    printf("%g\n", acosh(1.0));
-    printf("%g\n", atanh(0.0));
+    printf("%g\n", (double)(sinh(0.0)));
+    printf("%g\n", (double)(cosh(0.0)));
+    printf("%g\n", (double)(tanh(0.0)));
+    printf("%g\n", (double)(asinh(0.0)));
+    printf("%g\n", (double)(acosh(1.0)));
+    printf("%g\n", (double)(atanh(0.0)));
     return 0;
 }
