+++ test/cases/phase0/math/log-exp/expected.c
@@ -3,9 +3,9 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", log(1));
-    printf("%g\n", log2(8));
-    printf("%g\n", log10(100));
-    printf("%g\n", exp(0));
+    printf("%g\n", (double)(log(1)));
+    printf("%g\n", (double)(log2(8)));
+    printf("%g\n", (double)(log10(100)));
+    printf("%g\n", (double)(exp(0)));
     return 0;
 }
