+++ test/cases/phase0/math/log1p-expm1/expected.c
@@ -3,7 +3,7 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", log1p(0.0));
-    printf("%g\n", expm1(0.0));
+    printf("%g\n", (double)(log1p(0.0)));
+    printf("%g\n", (double)(expm1(0.0)));
     return 0;
 }
