+++ test/cases/phase0/math/pow-cbrt/expected.c
@@ -3,8 +3,8 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", pow(2, 10));
-    printf("%g\n", cbrt(27));
-    printf("%g\n", hypot(3, 4));
+    printf("%g\n", (double)(pow(2, 10)));
+    printf("%g\n", (double)(cbrt(27)));
+    printf("%g\n", (double)(hypot(3, 4)));
     return 0;
 }
