+++ test/cases/phase0/math/trig/expected.c
@@ -3,12 +3,12 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", sin(0));
-    printf("%g\n", cos(0));
-    printf("%g\n", tan(0));
-    printf("%g\n", asin(0));
-    printf("%g\n", acos(1));
-    printf("%g\n", atan(0));
-    printf("%g\n", atan2(1, 1));
+    printf("%g\n", (double)(sin(0)));
+    printf("%g\n", (double)(cos(0)));
+    printf("%g\n", (double)(tan(0)));
+    printf("%g\n", (double)(asin(0)));
+    printf("%g\n", (double)(acos(1)));
+    printf("%g\n", (double)(atan(0)));
+    printf("%g\n", (double)(atan2(1, 1)));
     return 0;
 }
