+++ test/cases/phase1/arithmetic/sub/expected.c
@@ -2,6 +2,6 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", 10 - 3);
+    printf("%g\n", (double)(10 - 3));
     return 0;
 }
