+++ test/cases/phase12/math/constants/expected.c
@@ -3,7 +3,7 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", M_PI);
-    printf("%g\n", M_E);
+    printf("%g\n", (double)(M_PI));
+    printf("%g\n", (double)(M_E));
     return 0;
 }
