+++ test/cases/phase12/math/max/expected.c
@@ -3,7 +3,7 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", fmax(3.0, 7.0));
-    printf("%g\n", fmax(-1.0, -5.0));
+    printf("%g\n", (double)(fmax(3.0, 7.0)));
+    printf("%g\n", (double)(fmax(-1.0, -5.0)));
     return 0;
 }
