+++ test/cases/phase12/math/trunc/expected.c
@@ -3,7 +3,7 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", trunc(3.9));
-    printf("%g\n", trunc(-3.9));
+    printf("%g\n", (double)(trunc(3.9)));
+    printf("%g\n", (double)(trunc(-3.9)));
     return 0;
 }
