+++ test/cases/phase2/number-type/param-number/expected.c
@@ -8,6 +8,6 @@ int main(void) {
     TSC_INIT();
     const double x = 1.5;
     const double y = 2.5;
-    printf("%g\n", add_f64_f64(x, y));
+    printf("%g\n", (double)(add_f64_f64(x, y)));
     return 0;
 }
