+++ test/cases/phase2/type-aliases/cross-compat/expected.c
@@ -7,6 +7,6 @@ int main(void) {
     TSC_INIT();
     const Pt1 a = { .x = 1.0, .y = 2.0 };
     const Pt2 b = *(const Pt2 *)&a;
-    printf("%g\n", b.x);
+    printf("%g\n", (double)(b.x));
     return 0;
 }
