+++ test/cases/phase2/type-aliases/struct-alias/expected.c
@@ -5,7 +5,7 @@ typedef struct { double x; double y; } Point;
 int main(void) {
     TSC_INIT();
     const Point p = { .x = 1.0, .y = 2.0 };
-    printf("%g\n", p.x);
-    printf("%g\n", p.y);
+    printf("%g\n", (double)(p.x));
+    printf("%g\n", (double)(p.y));
     return 0;
 }
