+++ test/cases/phase2/unknown/narrow-else/expected.c
@@ -37,7 +37,7 @@ int main(void) {
     TSC_INIT();
     tsc_unknown x = tsc_unknown_from_f64(42);
     if (x.type_id == 4) {
-        printf("%g\n", tsc_unknown_get_f64(&x) * 2);
+        printf("%g\n", (double)(tsc_unknown_get_f64(&x) * 2));
     } else {
         printf("%g\n", 0.0);
     }
