+++ test/cases/phase2/utility-types/record-literal/expected.c
@@ -5,6 +5,6 @@ typedef struct { double x; double y; } Coords;
 int main(void) {
     TSC_INIT();
     const Coords c = { .x = 1.0, .y = 2.0 };
-    printf("%g\n", c.x);
+    printf("%g\n", (double)(c.x));
     return 0;
 }
