+++ test/cases/phase2/widening/i32-to-f64/expected.c
@@ -4,6 +4,6 @@ int main(void) {
     TSC_INIT();
     const int32_t a = 7;
     const double b = a;
-    printf("%g\n", b);
+    printf("%g\n", (double)(b));
     return 0;
 }
