+++ test/cases/phase3/arrays/at/expected.c
@@ -6,7 +6,7 @@ int main(void) {
     TSC_INIT();
     double _arr_data_0[] = {10, 20, 30};
     const Array_f64 arr = (Array_f64){.data = _arr_data_0, .length = 3, .capacity = 3};
-    printf("%g\n", tsc_array_at_f64(arr, 0));
-    printf("%g\n", tsc_array_at_f64(arr, -1));
+    printf("%g\n", (double)(tsc_array_at_f64(arr, 0)));
+    printf("%g\n", (double)(tsc_array_at_f64(arr, -1)));
     return 0;
 }
