+++ test/cases/phase3/clone/deep-struct/expected.c
@@ -8,7 +8,7 @@ int main(void) {
     const Shape s = { .center = { .x = 1.0, .y = 2.0 }, .radius = 5.0 };
     const Shape s2 = (Shape)(s);
     s2.radius = 10.0;
-    printf("%g\n", s.radius);
-    printf("%g\n", s2.radius);
+    printf("%g\n", (double)(s.radius));
+    printf("%g\n", (double)(s2.radius));
     return 0;
 }
