+++ test/cases/phase3/type-conversion/f64-parse/expected.c
@@ -4,7 +4,7 @@ int main(void) {
     TSC_INIT();
     const String s = STR_LIT("3.14");
     const double n = tsc_parse_f64(s);
-    printf("%g\n", n);
+    printf("%g\n", (double)(n));
     tsc_string_release(s);
     return 0;
 }
