+++ test/cases/phase7/promise/then/expected.c
@@ -24,7 +24,7 @@ static void main_poll(main_state *self) {
             if (!self->_await_0._done) return;
             int32_t x = self->_await_0._result;
             const double y = x * 2;
-            printf("%g\n", y);
+            printf("%g\n", (double)(y));
             self->_done = true;
             return;
     }
